feat(types): generate frontend API types from Go wire structs
The frontend's API types were hand-written mirrors of the Go structs
and had drifted: Contact was missing how_met and profile_photo, marked
always-emitted has_pending_followup optional, invented a deleted_at
field the API never sends, UpdateContactRequest wrongly made full_name
optional, and the api-client typed error.details as an object where
the backend sends a string.

Pipeline: tygo (Go 1.24 tool directive) reads backend/tygo.yaml and
generates frontend/src/types/generated/ from the structs that actually
serialize the wire payloads — the swagger docs are stale Swagger 2.0
covering 2 of ~88 routes, so the Go structs are the source of truth.
Contact DTOs move to handlers/contact_dto.go (pure code motion; tstype
tags are generation metadata only). make api-types regenerates;
make api-types-check is a non-mutating drift gate wired into a new CI
api-types-drift job and the pre-push LINT lane.

frontend/src/types/contact.ts becomes a re-export of the generated
types, and api-client's envelope points at the generated
APIError/Meta. Type-level contract tests (expectTypeOf, enforced by
the tsc gate) were written first against the hand-written types and
failed on exactly the drift list above.

Adoption is incremental: further domains migrate by widening
tygo.yaml's include_files and re-pointing their types file.
