+++ .ai/pre-push.json
@@ -9,7 +9,8 @@
       { "name": "Frontend lint", "command": "cd frontend && bun run lint" },
       { "name": "Repo hygiene", "command": "bash scripts/hooks/repo-hygiene-check.sh" },
       { "name": "E2E test-map coverage", "command": "bash scripts/hooks/test-map-coverage-check.sh" },
-      { "name": "Spec lint", "command": "make spec-lint" }
+      { "name": "Spec lint", "command": "make spec-lint" },
+      { "name": "API types drift", "command": "make api-types-check" }
     ]
   },
   "swift": {
