+++ .ai/rules/core.md
@@ -204,6 +204,7 @@ Cross-cutting concerns that require checking multiple locations:
 | Scheduled job changes | Scheduler section in `.ai/guides/architecture.md` |
 | New sortable contact field | Shared ORDER BY ladder in ListContacts AND ListContactIDs (copies — keep in lockstep) + handler oneof + `SortField` type + `ContactListParams.sort` |
 | New contact list filter param | 3 unified SQL queries (ListContacts/ListContactIDs/CountContacts, shared WHERE) + repo params + handler query struct + frontend types + API client + contacts page + detail page listContext |
+| Wire DTO structs covered by `backend/tygo.yaml` (e.g. `handlers/contact_dto.go`, `api/response.go`) | Run `make api-types` and commit the regenerated `frontend/src/types/generated/*` — the CI api-types-drift job and the pre-push LINT lane (`make api-types-check`) fail on drift. New JSON payload structs for a covered domain go in the `*_dto.go` file, not the handler file |
 | Semantic column to table with dedup logic | All FindInWindow queries (add param), unique constraints, repository signatures, service call sites, `make sqlc`, regression test for false-positive case |
 | New build/run site for `crm-api` or `crm-admin`, or a new worker/type the binary needs | Build the PACKAGE (`./cmd/crm-api` / `./cmd/crm-admin`), NEVER a file (`cmd/crm-api/main.go`). New workers/types/adapters go in sibling `.go` files in `package main` (e.g. `cmd/crm-api/wire_*.go`) — do NOT inline into `main.go`. Build sites: `Makefile` (build/crm-admin/ci-build-backend/api-build/test-cadence-*), `.github/workflows/ci.yml`, `.github/workflows/build-images.yml`, `frontend/playwright.config.ts`, `scripts/start-backend.sh`, `scripts/dev-seed.sh`, `infra/install-systemd.sh`. Sole file-mode exception: `swag init -g cmd/crm-api/main.go` (`api-docs`). Confirm the ldflags stamp still lands (`build-images.yml` grep + `TestHealthEndpoint_StampedBuildInfo`). |
 | Path-filter groups (`path-filters.yml`) | CI dorny step (`.github/workflows/ci.yml` `changes` job) AND `scripts/hooks/pre-push` group-aware parser (`file_in_group` call sites + `check_swift`) AND the filter unit-test harness assertions |
