+++ .github/workflows/ci.yml
@@ -151,6 +151,28 @@ jobs:
       - name: Run spec linter
         run: make spec-lint
 
+  # Guards drift between the Go wire structs and the committed generated
+  # frontend types (backend/tygo.yaml → frontend/src/types/generated/).
+  # Gated on backend OR frontend: a backend DTO change regenerates differently,
+  # and a frontend hand-edit of a generated file diverges from the source.
+  # Deliberately minimal (no caches): tygo's dependency closure is small, so
+  # the cold path is seconds.
+  api-types-drift:
+    name: API Types Drift
+    runs-on: ubuntu-latest
+    needs: changes
+    if: ${{ needs.changes.outputs.backend == 'true' || needs.changes.outputs.frontend == 'true' }}
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+      - name: Set up Go
+        uses: actions/setup-go@v5
+        with:
+          go-version: '1.23'
+          cache: false
+      - name: Check generated API types are in sync
+        run: make api-types-check
+
   backend-tests:
     name: Backend Tests
     runs-on: ubuntu-latest
