+++ frontend/.prettierignore
@@ -14,6 +14,7 @@ node_modules/
 # Generated files
 *.min.js
 *.min.css
+src/types/generated/
 
 # Package manager
 package-lock.json
